LinkUrlDialog Refactor
Moving relevant code from CoreEditor to LinkUrlDialog

Broke some things need more work on the states
